Bug#37385923 routertest_integration_routing_sharing select_overlong times out
Post-push fix for broken windows build:
error C2672: 'std::min': no matching overloaded function found

std::min(sz, 16UL * 1024); is rejected on Win
std::min(sz, 16ULL * 1024); is rejected on macOS

so we must static_cast it to size_t;

+++ router/src/routing/src/channel.cc
@@ -239,7 +239,7 @@ stdx::expected<size_t, std::error_code> Channel::read_to_plain(size_t sz) {
     //   1k |  2100ms  | 256ms
     //  16k |  1966ms  | 219ms
     // 256k | *6509ms* | 189ms
-    const size_t to_read = std::min(sz, 16UL * 1024);
+    const size_t to_read = std::min(sz, static_cast<size_t>(16UL * 1024));
 
     auto dyn_buf = net::dynamic_buffer(plain_buf);
